comment order check for testing

+++ src/modules/review/review.service.ts
@@ -1,25 +1,27 @@
-import { Request } from 'express'
-import { IReview } from '../../models'
 import Product from '../../models/product.model'
-import { ReviewRepository } from '../../repositories'
+import { OrderRepository, ReviewRepository } from '../../repositories'
+import { ErrorMessages } from '../../shared/constants/errorMessages'
 import { BadRequestError, NotFoundError } from '../../shared/errors/errors'
 import { ID } from '../../shared/types'
 import { ProductService } from '../product/product.service'
+import { CreateReviewDTO, UpdateReviewDTO } from './review.dtos'
 
 export class ReviewService {
     private reviewRepository: ReviewRepository
     private productService: ProductService
+    private orderRepository: OrderRepository
 
     constructor() {
         this.reviewRepository = new ReviewRepository()
         this.productService = new ProductService()
+        this.orderRepository = new OrderRepository()
     }
 
     async getAllReviews(filter: Record<string, any> = {}) {
         return this.reviewRepository.findAll({
             filter,
             populate: [
-                { path: 'user', select: '_id fullname' },
+                { path: 'createdBy', select: '_id fullname' },
                 { path: 'product', select: '_id name' },
             ],
         })
@@ -29,27 +31,40 @@ export class ReviewService {
         return this.reviewRepository.findById(id)
     }
 
-    async createReview(req: Request, product: ID, user: ID) {
-        const { review, rating } = req.body
+    async createReview(body: CreateReviewDTO, product: ID, user: ID) {
+        const { review, rating } = body
 
         const productFound = await Product.findById(product)
         if (!productFound) {
-            throw new NotFoundError('Product not found')
+            throw new NotFoundError(ErrorMessages.PRODUCT_NOT_FOUND)
         }
 
+        // const userOrder = await this.orderRepository.findOne({
+        //     user: user,
+        //     product: product,
+        // })
+        // if (!userOrder) {
+        //     throw new BadRequestError(
+        //         ErrorMessages.YOU_HAVE_NOT_PURCHASED_THIS_PRODUCT,
+        //     )
+        // }
+
         const existingReview = await this.reviewRepository.exists({
-            user,
+            createdBy: user,
             product,
         })
         if (existingReview) {
-            throw new BadRequestError('You have already reviewed this product.')
+            throw new BadRequestError(
+                ErrorMessages.YOU_HAVE_ALREADY_REVIEWED_THIS_PRODUCT,
+            )
         }
 
         const newReview = await this.reviewRepository.create({
-            user,
+            createdBy: user,
             product,
             review,
             rating,
+            order: user,
         })
 
         await this.calcAverageRatings(product)
@@ -57,20 +72,33 @@ export class ReviewService {
         return newReview
     }
 
-    async updateReview(id: ID, updateData: Partial<IReview>) {
-        const updatedReview = await this.reviewRepository.updateById(
-            id,
-            updateData,
-        )
+    async updateReview(id: ID, userId: ID, body: UpdateReviewDTO) {
+        const review = await this.reviewRepository.findOne({
+            _id: id,
+            createdBy: userId,
+        })
 
+        if (!review) {
+            throw new NotFoundError(ErrorMessages.REVIEW_NOT_FOUND)
+        }
+
+        const updatedReview = await this.reviewRepository.updateById(id, body)
         await this.calcAverageRatings(updatedReview.product)
 
         return updatedReview
     }
 
-    async deleteReview(id: ID) {
-        const deletedReview = await this.reviewRepository.deleteById(id)
+    async deleteReview(id: ID, userId: ID) {
+        const review = await this.reviewRepository.findOne({
+            _id: id,
+            createdBy: userId,
+        })
 
+        if (!review) {
+            throw new NotFoundError(ErrorMessages.REVIEW_NOT_FOUND)
+        }
+
+        const deletedReview = await this.reviewRepository.deleteById(id)
         await this.calcAverageRatings(deletedReview.product)
 
         return deletedReview
@@ -79,16 +107,11 @@ export class ReviewService {
     async calcAverageRatings(productId: ID) {
         const stats = await this.reviewRepository.calcAverageRatings(productId)
 
-        if (stats.length > 0) {
-            await this.productService.updateProduct(productId, {
-                feedbacks: stats[0].nRating,
-                averageRating: stats[0].avgRating,
-            })
-        } else {
-            await this.productService.updateProduct(productId, {
-                feedbacks: 0,
-                averageRating: 0,
-            })
+        const updateData = {
+            feedbacks: stats.length > 0 ? stats[0].nRating : 0,
+            averageRating: stats.length > 0 ? stats[0].avgRating : 0,
         }
+
+        await this.productService.updateProductRatings(productId, updateData)
     }
 }
